feat/delete complete

+++ src/handler/content.ts
@@ -8,6 +8,7 @@ import { getOEmbedInfo } from "../utils/oembed";
 import mapToDto from "../utils/content.mapper";
 import contentMapper from "../utils/content.mapper";
 import { IUpdateContent } from "../repositories";
+import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
 
 export default class ContentHandler implements IContentHandler {
   private repo: IContentRepository;
@@ -88,8 +89,35 @@ export default class ContentHandler implements IContentHandler {
   };
 
   public deleteById: IContentHandler["deleteById"] = async (req, res) => {
-    const result = await this.repo.delete(Number(req.params.id));
+    try {
+      if (isNaN(Number(req.params.id)))
+        return res.status(400).json({ message: "id is invalid" });
 
-    return res.status(200).json(result).end();
+      const {
+        postedBy: { id: ownerId },
+      } = await this.repo.getById(Number(req.params.id));
+      if (ownerId !== res.locals.user.id)
+        return res
+          .status(403)
+          .json({ message: "You're not the owner of this content!" });
+
+      const result = await this.repo.delete(Number(req.params.id));
+
+      return res.status(200).json(result).end();
+    } catch (error) {
+      console.error(error);
+      if (
+        error instanceof PrismaClientKnownRequestError &&
+        error.code === "P2025"
+      )
+        return res.status(410).json({ message: "content not found" }).end();
+
+      if (error instanceof TypeError)
+        return res.status(400).json({ message: error.message }).end();
+
+      return res.status(500).json({
+        message: `Internal Server Error`,
+      });
+    }
   };
 }
